+++ src/index.js
@@ -1,6 +1,10 @@
+import "./index.scss"
 import calc from './calculator'
-import  sayHello  from './hello'
+import sayHello from './hello'
 
-console.log(calc(3, 2, 'subs'))
+alert('mafe soupe kandia domoda, tiep bou honk tiep bou weikh bou am soule')
+// alert('css ', css)
+const a = calc(3, 2, 'subs')
+alert(a)
 
-sayHello('Elhadji')
+// sayHello('Elhadji')