+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # webpack configuration
     - webpack conf
     - babel conf
+    - sass loader
+    - css loader
+    - style loader